// src/app/core/models/Property.ts
import { NumberSymbol } from "@angular/common/src/i18n/locale_data_api";

export class Property {

    PropertyId: number;
    ListingID: any;
    PropertyName: string;
    Address: string;
    Country: String;
    City: string;
    State: string;
    ZipCode: string;
    GeneralUse: string;
    SpecificUse: string;
    BuildingSize: number;
    BuildingSizeSM: number;
    GrossBuildingArea: number;
    YearBuilt: number;
    YearRenovated: number;
    NumberOfFloor: number;
    SizeSource: string;
    MixedUseAllocation: string;
    ConstructionType: string;
    LotSizeAcres: number;
    Zoning: string;
    ParkingSpaces: number;
    ParkingRatio: number;
    AdditionalUse1: string;
    AdditionalUse2: string;
    Condo: string;
    OwnerOccupied: string;
    Tenancy: string;
    Class: string;
    ClearHeight: number;
    HVACType: string;
    ColumnSpacing: string;
    SprinklerType: string;
    ADAAccessible: string;
    Quardrant: number;
    ParkComplexName: string;
    ParcelInformation: string;
    Latitude: number;
    Longitude: number;
    Metro: string;
    Market: string;
    SubMarket: string;
    PropertyManagerName: string;
    PropertyManagerPhone: number;
    PropertyManagerEmail: string;
    PropertyManagerWebsite: string;
    PropertyManagerAddress: string;
    ImageUrl: string;
    ConstructionStatus: string;
    RetailFrontage: string;
    IsMixedUse: string;
    TotalOfficeSF: number;
    IsCraneServed: string;
    IsRailServed: string;
    NumberOfDockDoor: number;
    NumberOfGradeLevelkDoors: number;
    NumberOfTruckkDoors: number;
    Phase: number;
    Volts: number;
    Amps: number;
    NumberOfOffices: number;
    NumberOfRestRooms: number;
    YardFencedStatus: string;
    YardFencedArea: number;
    YardUnFencedStatus: string;
    YardUnFencedArea: number;
    YardPavedStatus: string;
    YardPavedArea: number;
    CommonAreaFactor: number;
    RetailSF: number;
    MinFloorSize: number;
    MaxFloorSize: number;
    AvgFloorSize: number;
    NumberOfOfficeFloors: number;
    NumberOfAnchors: number;
    AnchorSf: number;
    BayDepthWidth; string;
    VerifiedStatus: string;
    TrafficCount: number;
    ListingType: string;
    // MinAvailableSpace: number;
    // MaxAvailableSpace: number;
    // MinLeaseRate: number;
    // MaxLeaseRate: number;
    // MinBuildingSize: number;
    // MaxBuildingSize: number;
    District: String
    BuildingClass: string;
    ParcelNumber: string;
    TypicalFloorArea: number;
    NumberOfCommercialUnits: string;
    NumberOfElevators: number;
    NumberOfDockDoors: number;
    NumberOfDriveInDoors: number;
    NumberOfTruckwellDoors: number;
    Rail: string;
    LotSizeSM: number;
    Sprinkler: string;
    Notes: string;
    PropertyImageUrl1: string;
    PropertyImageUrl2: string;
    MapImageUrl: string;
    PropertyMedia: Array<any>;
    ListingMedia: Array<any>;
    SuiteMedia: Array<any>;

    AvailableSpace: number;
    LeaseRate: number;
    SalePrice: number;
    DateAvailable: string;
    ExcludeNegotiableRate: string;
    TermType: string;
    OccupancyPercent: number;
    isSelected: boolean = false;
    TotalAvailable: number;
    TotalAvailableSM: number;
    DisplayOrder: number;

    MainPhotoUrl: string;
    PropertyComments: string;
    MarketName: string;

    Amenities: string;
    BuildingComments: string;
    CoreFactor: string;



    // new
    County: string;
    EastWestStreet: any = null;
    NorthSouthStreet: any = null;
    Floors: any;
    LotSizeSourceName: any;
    TotalAnchorSF: any;
    IsVented: any;
    HasResCoveredParking: any;
    HasResSurfaceParking: any;
    DockHigh: any;
    GradeLevelDriveIn: any;
    Truckwell: any;
    RailServed: any;
    CraneServed: any;
    HasYard: any;
    HasYardFenced: any;
    HasPavedYard: any;
    SpecificUseId: any;
    GeneralUseId: any;
}